+++ tests/test/__init__.py
@@ -1,5 +1,7 @@
 """Root class for testing."""
 from django.test import TestCase, Client, RequestFactory
+from django.contrib.admin import site
+from django.urls import reverse
 
 
 class TestBase(TestCase):
@@ -20,6 +22,19 @@ class TestBase(TestCase):
           email='mail@example.com',
           password=self.admin_pass
         )
+
+        from example.admin import Admin
+        from example.models import Log
+
+        self.modeladmin = Admin(Log, site)
+        self.url = reverse('admin:example_log_changelist')
+        self.queryset = Log.objects.all()
+
+    def admin_get(self, params):
+        """Get request from admin."""
+        request = self.request_factory.get(self.url, params)
+        request.user = self.admin
+        return request
 
     def login_admin(self):
         """Login as admin."""
